--- doc/GBAnomalyTrackerCollection.json
@@ -1,229 +1,0 @@
-{
-	"info": {
-		"_postman_id": "393a289e-bec7-4c48-bb6d-7abc93bf4a63",
-		"name": "GB Anomaly Tracker API",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
-	},
-	"item": [
-		{
-			"name": "Criar Detector",
-			"request": {
-				"method": "POST",
-				"header": [
-					{
-						"key": "Content-Type",
-						"value": "application/json"
-					}
-				],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"detectorName\": \"Detector de volume de vendas\",\n    \"metricName\": \"Volume de Vendas\",\n    \"alertVariation\": 10,\n    \"alertIntervalMinutes\": 1,\n    \"alertMessage\": \"Foi detectada uma variação de 10% abaixo do volume de vendas padrão\",\n    \"alertChannel\": [\"Slack\", \"Telegram\"],\n    \"direction\": \"abaixo\",\n    \"action\": \"Acionar o time de vendas\",\n    \"objective\": \"Monitorar e identificar quedas significativas no volume de vendas\",\n    \"aggregationMethod\": \"sum\"\n}"
-				},
-				"url": {
-					"raw": "{{baseUrl}}/detectors",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"detectors"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Obter Todos os Detectores",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "{{baseUrl}}/detectors",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"detectors"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Obter Detector por ID",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "{{baseUrl}}/detectors/1",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"detectors",
-						"1"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Atualizar Detector",
-			"request": {
-				"method": "PUT",
-				"header": [
-					{
-						"key": "Content-Type",
-						"value": "application/json"
-					}
-				],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"detectorName\": \"Detector de volume de vendas\",\n    \"metricName\": \"Volume de Vendas\",\n    \"alertVariation\": 20,\n    \"alertIntervalMinutes\": 30,\n    \"alertMessage\": \"Alerta! Variação detectada\",\n    \"alertChannel\": [\"Slack\", \"Telegram\"],\n    \"action\": \"Acionar o time de vendas\",\n    \"objective\": \"Monitorar e identificar quedas significativas no volume de vendas\"\n}"
-				},
-				"url": {
-					"raw": "{{baseUrl}}/detectors/1",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"detectors",
-						"1"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Deletar Detector",
-			"request": {
-				"method": "DELETE",
-				"header": [],
-				"url": {
-					"raw": "{{baseUrl}}/detectors/1",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"detectors",
-						"1"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Registrar Métrica",
-			"request": {
-				"method": "POST",
-				"header": [
-					{
-						"key": "Content-Type",
-						"value": "application/json"
-					}
-				],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"metricName\": \"Volume de Vendas\",\n    \"value\": 1\n}"
-				},
-				"url": {
-					"raw": "{{baseUrl}}/metrics",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"metrics"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Registrar Métrica Com Parametro",
-			"request": {
-				"method": "POST",
-				"header": [
-					{
-						"key": "Content-Type",
-						"value": "application/json"
-					}
-				],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"metricName\": \"Valor produto\",\n    \"value\": 2,\n    \"params\": [\n        {\n            \"key\": \"id\",\n            \"value\": \"15\"\n        }\n    ]\n}"
-				},
-				"url": {
-					"raw": "{{baseUrl}}/metrics",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"metrics"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Listar Métricas",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "{{baseUrl}}/metrics",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"metrics"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Obter Métricas Agrupadas por Minuto",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "{{baseUrl}}/detectors/1/telemetry",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"detectors",
-						"1",
-						"telemetry"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Criar Detector Parametrizado",
-			"request": {
-				"method": "POST",
-				"header": [
-					{
-						"key": "Content-Type",
-						"value": "application/json"
-					}
-				],
-				"body": {
-					"mode": "raw",
-					"raw": "{\n    \"detectorName\": \"Detector de volume de vendas de IPHONE\",\n    \"metricName\": \"Volume de Vendas\",\n    \"alertVariation\": 10,\n    \"alertIntervalMinutes\": 1,\n    \"alertMessage\": \"Foi detectada uma variação de 10% abaixo do volume de vendas de IPHONE\",\n    \"alertChannel\": [\"Slack\", \"Telegram\"],\n    \"direction\": \"abaixo\",\n    \"action\": \"Acionar o time de vendas de IPHONE\",\n    \"objective\": \"Monitorar e identificar quedas significativas no volume de vendas de IPHONE\",\n    \"aggregationMethod\": \"sum\",\n    \"params\": [\n        {\n            \"key\": \"produto\",\n            \"value\": \"iphone\"\n        }\n    ]\n}"
-				},
-				"url": {
-					"raw": "{{baseUrl}}/detectors",
-					"host": [
-						"{{baseUrl}}"
-					],
-					"path": [
-						"detectors"
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
