+++ doc/exemplo-html-com-grafico-dinamico.html
@@ -13,7 +13,7 @@
 </head>
 <body>
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <a class="navbar-brand" href="#">GB Anomaly Tracker</a>
+        <a class="navbar-brand" href="#">TM Anomaly Tracker</a>
         <div class="collapse navbar-collapse">
             <ul class="navbar-nav">
                 <li class="nav-item">
