+++ doc/simulador-metricas.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>Simulador de envio de telemetria GB Anomaly Tracker</title>
+    <title>Simulador de envio de telemetria TM Anomaly Tracker</title>
     <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/material-design-icons/3.0.1/iconfont/material-icons.min.css">
@@ -28,7 +28,7 @@
 </head>
 <body>
     <div class="container">
-        <h1>Simulador de envio de telemetria GB Anomaly Tracker</h1>
+        <h1>Simulador de envio de telemetria TM Anomaly Tracker</h1>
 
         <div class="input-field">
             <label for="metricName">Nome da Métrica:</label>
