+++ pom.xml
@@ -8,11 +8,11 @@
 		<version>3.1.1</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>br.com.gbanomalytracker</groupId>
-	<artifactId>gb-anomaly-tracker-api</artifactId>
+	<groupId>br.com.tmanomalytracker</groupId>
+	<artifactId>tm-anomaly-tracker-api</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>gb-anomaly-tracker-api</name>
-	<description>GB Anomaly Tracker API</description>
+	<name>tm-anomaly-tracker-api</name>
+	<description>TM Anomaly Tracker API</description>
 	<properties>
 		<java.version>17</java.version>
 		<kotlin.version>1.8.22</kotlin.version>
