--- src/main/kotlin/br/com/gbanomalytracker/client/NotificationClient.kt
@@ -1,7 +1,0 @@
-package br.com.gbanomalytracker.client
-
-interface NotificationClient {
-    fun sendNotification(message: String)
-
-    fun getChannel(): String
-}
