--- src/main/kotlin/br/com/gbanomalytracker/dto/MetricPeriodResponse.kt
@@ -1,9 +1,0 @@
-package br.com.gbanomalytracker.dto
-
-import java.time.LocalDateTime
-
-data class MetricPeriodResponse(
-    val startTime: LocalDateTime,
-    val endTime: LocalDateTime,
-    val valueSum: Double,
-)
