--- src/main/kotlin/br/com/gbanomalytracker/dto/TelemetryAggregate.kt
@@ -1,8 +1,0 @@
-package br.com.gbanomalytracker.dto
-
-import java.time.LocalDateTime
-
-data class TelemetryAggregate(
-    val timestamp: LocalDateTime,
-    val value: Double
-)
