--- src/main/kotlin/br/com/gbanomalytracker/repository/AnomalyAnalysisResultRepository.kt
@@ -1,8 +1,0 @@
-package br.com.gbanomalytracker.repository
-
-import br.com.gbanomalytracker.entity.AnomalyAnalysisResult
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface AnomalyAnalysisResultRepository : JpaRepository<AnomalyAnalysisResult, Long>
