+++ src/main/kotlin/br/com/tmanomalytracker/client/NotificationClient.kt
@@ -1,0 +1,7 @@
+package br.com.tmanomalytracker.client
+
+interface NotificationClient {
+    fun sendNotification(message: String)
+
+    fun getChannel(): String
+}
