+++ src/main/kotlin/br/com/tmanomalytracker/dto/MetricPeriodResponse.kt
@@ -1,0 +1,9 @@
+package br.com.tmanomalytracker.dto
+
+import java.time.LocalDateTime
+
+data class MetricPeriodResponse(
+    val startTime: LocalDateTime,
+    val endTime: LocalDateTime,
+    val valueSum: Double,
+)
