+++ src/main/kotlin/br/com/tmanomalytracker/dto/TelemetryAggregate.kt
@@ -1,0 +1,8 @@
+package br.com.tmanomalytracker.dto
+
+import java.time.LocalDateTime
+
+data class TelemetryAggregate(
+    val timestamp: LocalDateTime,
+    val value: Double
+)
