+++ src/main/resources/application.properties
@@ -1,13 +1,13 @@
-# Configurações do JPA
-spring.datasource.url=jdbc:postgresql://localhost:5432/gb-anomaly-tracker-db
-spring.datasource.username=gb-anomaly-tracker-db-user
+spring.datasource.url=jdbc:postgresql://localhost:5432/tm-anomaly-tracker-db
+spring.datasource.username=tm-anomaly-tracker-db-user
+spring.datasource.password=tm-anomaly-tracker-db-pass
 spring.datasource.password=gb-anomaly-tracker-db-pass
 
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
 
-# Configurações do Open API e Swagger Docs
+# ConfiguraÃ§Ãµes do Open API e Swagger Docs
 springdoc.api-docs.path=/api-docs
 springdoc.swagger-ui.path=/swagger-ui.html
 springdoc.swagger-ui.enabled=true
