--- src/test/kotlin/br/com/gbanomalytracker/GbAnomalyTrackerApiApplicationTests.kt
@@ -1,13 +1,0 @@
-package br.com.gbanomalytracker
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class GbAnomalyTrackerApiApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
