+++ src/test/kotlin/br/com/tmanomalytracker/TmAnomalyTrackerApiApplicationTests.kt
@@ -1,0 +1,13 @@
+package br.com.tmanomalytracker
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class TmAnomalyTrackerApiApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+
+}
